Working on bot logic

+++ Entities/WhitelistedItem.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; }
         public string Image { get; set; }
         public int MaxQuantity { get; set; }
-        public float PriceMultiplier { get; set; }
+        public decimal PriceMultiplier { get; set; }
         public int AccountId { get; set; }
         public Account Account { get; set; }
     }
