+++ Models/WhitelistedItems/CreateRequest.cs
@@ -7,7 +7,7 @@ namespace WebApi.Models.WhitelistedItems
         [Required]
         public string Name { get; set; }
         public int MaxQuantity { get; set; }
-        public float PriceMultiplier { get; set; }
+        public decimal PriceMultiplier { get; set; }
         public int AccountId { get; set; }
     }
 }
