+++ Models/WhitelistedItems/UpdateRequest.cs
@@ -3,6 +3,6 @@ namespace WebApi.Models.WhitelistedItems
     public class UpdateRequest
     {
         public int MaxQuantity { get; set; }
-        public float PriceMultiplier { get; set; }
+        public decimal PriceMultiplier { get; set; }
     }
 }
