+++ Models/WhitelistedItems/WhitelistedItemResponse.cs
@@ -5,7 +5,7 @@ namespace WebApi.Models.WhitelistedItems
         public int Id { get; set; }
         public string Name { get; set; }
         public int MaxQuantity { get; set; }
-        public float PriceMultiplier { get; set; }
+        public decimal PriceMultiplier { get; set; }
         public int AccountId { get; set; }
     }
 }
